container: fix builds on Docker Hub
Builds on Docker Hub are using some internal cache/proxy for CentOS
repositories which either broken or does not work well with modular
RPMs. Trying to install packages from `virt` module build fails on
downloading the RPMs. Easiest fix is to disable the `virt` module
because we don't need it. All the packages we need are in the Advanced
Virt repos we have configured from oVirt.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>