container: extract virtio-win.iso content
virt-v2v can use either the /usr/share/virtio-win directory or the ISO
as source for drivers/tools, not both. By default the ISO is picked if
it exists in the file system. This makes it tricky for us because we
store Linux qemu-ga RPMs only in the directory and the directory does
not contain all the drivers (only limited subset). Without any fall-back
mechanism in virt-v2v the best solution is to extract the ISO content
and remove it.

Signed-off-by: Tomáš Golembiovský <tgolembi@redhat.com>